+++ cedar-policy/src/api.rs
@@ -2217,6 +2217,43 @@ impl PolicySet {
         Ok(est)
     }
 
+    /// Get the human-readable Cedar syntax representation of this policy set.
+    /// This function is primarily intended for rendering JSON policies in the
+    /// human-readable syntax, but it will also return the original policy text
+    /// (though possibly re-ordering policies withing the policy set) when the
+    /// policy-set contains policies parsed from the human-readable syntax.
+    ///
+    /// This will return `None` if there are any linked policies in the policy
+    /// set because they cannot be directly rendered in Cedar syntax. It also
+    /// cannot record policy ids because these cannot be specified in the Cedar
+    /// syntax. The policies may be reordered, so parsing the resulting string
+    /// with [`PolicySet::from_str`] is likely to yield different policy id
+    /// assignments. For these reasons you should prefer serializing as JSON and
+    /// only using this function to obtain a representation to display to human
+    /// users.
+    ///
+    /// This function does not format the policy according to any particular
+    /// rules.  Policy formatting can be done through the Cedar policy CLI or
+    /// the `cedar-policy-formatter` crate.
+    pub fn to_cedar(&self) -> Option<String> {
+        let policies = self
+            .policies
+            .values()
+            // We'd like to print policies in a deterministic order, so we sort
+            // before printing, hoping that the size of policy sets is fairly
+            // small.
+            .sorted_by_key(|p| AsRef::<str>::as_ref(p.id()))
+            .map(|p| p.to_cedar())
+            .collect::<Option<Vec<_>>>()?;
+        let templates = self
+            .templates
+            .values()
+            .sorted_by_key(|t| AsRef::<str>::as_ref(t.id()))
+            .map(|t| t.to_cedar());
+
+        Some(policies.into_iter().chain(templates).join("\n\n"))
+    }
+
     /// Create a fresh empty `PolicySet`
     pub fn new() -> Self {
         Self {
@@ -2901,6 +2938,21 @@ impl Template {
         serde_json::to_value(est).map_err(Into::into)
     }
 
+    /// Get the human-readable Cedar syntax representation of this template.
+    /// This function is primarily intended for rendering JSON policies in the
+    /// human-readable syntax, but it will also return the original policy text
+    /// when given a policy parsed from the human-readable syntax.
+    ///
+    /// It also does not format the policy according to any particular rules.
+    /// Policy formatting can be done through the Cedar policy CLI or
+    /// the `cedar-policy-formatter` crate.
+    pub fn to_cedar(&self) -> String {
+        match &self.lossless {
+            LosslessPolicy::Est(_) => self.ast.to_string(),
+            LosslessPolicy::Text { text, .. } => text.clone(),
+        }
+    }
+
     /// Get valid [`RequestEnv`]s.
     /// A [`RequestEnv`] is valid when the template type checks w.r.t requests
     /// that satisfy it.
@@ -3378,6 +3430,33 @@ impl Policy {
     pub fn to_json(&self) -> Result<serde_json::Value, PolicyToJsonError> {
         let est = self.lossless.est()?;
         serde_json::to_value(est).map_err(Into::into)
+    }
+
+    /// Get the human-readable Cedar syntax representation of this policy. This
+    /// function is primarily intended for rendering JSON policies in the
+    /// human-readable syntax, but it will also return the original policy text
+    /// when given a policy parsed from the human-readable syntax.
+    ///
+    /// It will return `None` for linked policies because they cannot be
+    /// directly rendered in Cedar syntax. You can instead render the unlinked
+    /// template if you do not need to preserve links. If serializing links is
+    /// important, then you will need to serializing the whole policy set
+    /// containing the template and link to JSON.
+    ///
+    /// It also does not format the policy according to any particular rules.
+    /// Policy formatting can be done through the Cedar policy CLI or
+    /// the `cedar-policy-formatter` crate.
+    pub fn to_cedar(&self) -> Option<String> {
+        match &self.lossless {
+            LosslessPolicy::Est(_) => Some(self.ast.to_string()),
+            LosslessPolicy::Text { text, slots } => {
+                if slots.is_empty() {
+                    Some(text.clone())
+                } else {
+                    None
+                }
+            }
+        }
     }
 
     /// Get all the unknown entities from the policy